# backend/swagger.yaml
openapi: "3.0.0"
info:
  title: McMaster CanIDS
  version: 2.0.0
paths:
...
  /api/user/add:
    post:
      summary: Create new user
      description: |
        Create a user account on someone's behalf.
        
        Restrictions:
          - `superuser` can create all types of accounts for all groups.
          - `admin` can create `admin` and `standard` accounts for same `group` UUID.
          - `standard` can not create accounts.
          
        If an account is successfully created, the user will be emailed with instructions to activate the account and set a password.
      tags:
      - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: |
                    Name of user
                uuid:
                  type: string
                  description: |
                    Email address of user
                class:
                  type: string
                  description: |
                    Type of user account: `standard`, `admin`, `superuser`
                group:
                  type: string
                  description: |
                    Group UUID
              required:
                - name
                - uuid
                - class
                - group
            example: {
              "name": "Tanner Ryan",
              "uuid": "ryant3@mcmaster.ca",
              "class": "admin",
              "group": "175dd231-f547-4342-a33e-c45d5e99aa7e"
            }
      responses:
        '200':
          description: |
            Account successfully created. User emailed with activation instructions.
          content:
            application/json:
              example: {
                "success": true,
                "message": "The user account has been successfully created. The user has been emailed to complete account activation."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Admin users can not add superusers."
              }
        '500':
          description: |
            Internal server error
...
  /api/group/add:
    post:
      summary: Create new group
      description: |
        Create a new group.
        
        Restrictions:
          - `superuser` is the only class of accounts allowed to create groups.

      tags:
      - Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: |
                    Name of user
                authorized:
                  type: array
                  description: |
                    List of authorized assets
                  items:
                    type: string
              required:
                - name
                - authorized
            example: {
              "name": "McMaster University",
              "authorized": [
                "mcmaster_001",
                "mcmaster_002"
              ],
            }
      responses:
        '200':
          description: |
            Group successfully created.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Group successfully created."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Only a superuser can update a group."
              }
        '500':
          description: |
            Internal server error
  /api/group/update:
    post:
      summary: Update existing group
      description: |
        Update existing group.
        
        Restrictions:
          - `superuser` is the only class of accounts allowed to update groups.

      tags:
      - Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid:
                  type: string
                  description: |
                    UUID of group to update
                name:
                  type: string
                  description: |
                    Updated or existing group name
                authorized:
                  type: array
                  description: |
                    Updated or existing group assets
                  items:
                    type: string
              required:
                - uuid
                - name
                - authorized
            example: {
              "uuid": "175dd231-f547-4342-a33e-c45d5e99aa7e",
              "name": "McMaster University",
              "authorized": [
                "asset1",
                "asset2"
              ]
            }
      responses:
        '200':
          description: |
            Group successfully deleted.
          content:
            application/json:
              example: {
                "success": true,
                "message": "Successfully deleted group."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Only a superuser can delete a group."
              }
        '500':
          description: |
            Internal server error
...
  /api/view/add:
    post:
      summary: Create new view
      description: |
        Create a view group.
        
        Restrictions:
          - `admin` can create views for their own group with authorized assets registered to their group.
          - `superuser` can create views for any group with any authorized assets.
  
        Available view classes:
          - `line`
          - `bar`
          - `pie`
          - `table`
      tags:
      - View
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                group:
                  type: string
                  description: |
                    Group that view belongs to
                authorized:
                  type: string
                  description: |
                    Authorized asset to populate view with (source index)
                name:
                  type: string
                  description: |
                    Name of view
                class:
                  type: string
                  description: |
                    Type of view: `line`, `bar`, `pie`, `table`
                field:
                  type: string
                  description: |
                    Data field to use for view
                fieldName:
                  type: string
                  description: |
                    Common field name
              required:
                - group
                - authorized
                - name
                - class
                - field
                - fieldName
            example: {
              "group": "6a7d3a03-5c06-4620-8325-583c8136b17f",
              "authorized": "mcmaster_001",
              "name": "JHE Incoming Packets",
              "class": "line",
              "field": "packets_incoming",
              "fieldName": "Packets"
            }
      responses:
        '200':
          description: |
            View successfully created.
          content:
            application/json:
              example: {
                "success": true,
                "message": "View successfully created."
              }
        '400':
          description: |
            Request parameters are not valid.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Bad request format."
              }
        '401':
          description: |
            User is not authenticated
        '403':
          description: |
            User does not have permissions to perform requested actions.
          content:
            application/json:
              example: {
                "success": false,
                "message": "Admin users can not add foreign views."
              }
        '500':
          description: |
            Internal server error
...